+++ certbot-apache/certbot_apache/_internal/override_centos.py
@@ -62,7 +62,7 @@ class CentOSConfigurator(configurator.ApacheConfigurator):
         # https://github.com/certbot/certbot/issues/9481.
         if not rhel_derived:
             return False
-        at_least_v9 = util.LooseVersion(os_version).try_risky_greater_equal(util.LooseVersion('9'))
+        at_least_v9 = util.parse_loose_version(os_version) >= util.parse_loose_version('9')
         return at_least_v9
 
     def _override_cmds(self) -> None: