+++ certbot/CHANGELOG.md
@@ -6,12 +6,11 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-*
+* Add `certbot.util.LooseVersion` class. See [GH #9489](https://github.com/certbot/certbot/issues/9489).
 
 ### Changed
 
-* Deprecated `certbot.util.parse_loose_version`. Replaced with `certbot.util.LooseVersion` class.
-  See [GH #9489](https://github.com/certbot/certbot/issues/9489).
+*
 
 ### Fixed